+++ example/lib/main.dart
@@ -3,6 +3,7 @@ import 'dart:async';
 
 import 'package:flutter/services.dart';
 import 'package:tc_consent_plugin/tc_consent.dart';
+import 'package:tc_consent_plugin/tc_consent_api.dart';
 
 void main() {
   runApp(const MyApp());
@@ -43,6 +44,7 @@ class _MyAppState extends State<MyApp> {
               children: [
                 buildTextButton('Accept All Consent', consent.acceptAllConsent ),
                 buildTextButton('Refuse all consent', consent.refuseAllConsent ),
+                buildTextButton('Block privacy Center ',() => {consent.blockIOSPrivacyCenterDropOut = true} ),
                 buildTextButton('Show Privacy Center', consent.showPrivacyCenter ),
                 buildTextButton('Call use AC String',() {consent.useACString(true);} ),
                 buildTextButton('Init with Custom Privacy Center',() {consent.initWithCustomPCM(SITE_ID, PRIVACY_ID);} ),
@@ -62,6 +64,52 @@ class _MyAppState extends State<MyApp> {
                 buildTextButton('getConsentAsJson',() async {
                   var savedConsent = await consent.getConsentAsJson();
                 }),
+                buildTextButton('shouldDisplayPrivacyCenter',() async {
+                  var val = await TCConsentAPI.shouldDisplayPrivacyCenter();
+                  print("shouldDisplayPrivacyCenter = ${val}");
+                }),
+                buildTextButton('getAllAcceptedConsent',() async {
+                  var val = await TCConsentAPI.getAllAcceptedConsent();
+                  print("getAllAcceptedConsent = ${val}");
+                }),
+                // buildTextButton('getAcceptedGoogleVendors',() async {
+                //   var val = await TCConsentAPI.getAcceptedGoogleVendors();
+                //   print("getAcceptedGoogleVendors = ${val}");
+                // }),
+                buildTextButton('getAcceptedCategories',() async {
+                  var val = await TCConsentAPI.getAcceptedCategories();
+                  print("getAcceptedCategories = ${val}");
+                }),
+                buildTextButton('isIABSpecialFeatureAccepted',() async {
+                  var val = await TCConsentAPI.isIABSpecialFeatureAccepted(1);
+                  print("isIABSpecialFeatureAccepted = ${val}");
+                }),
+                buildTextButton('isIABVendorAccepted',() async {
+                  var val = await TCConsentAPI.isIABVendorAccepted(1);
+                  print("isIABVendorAccepted = ${val}");
+                }),
+                buildTextButton('isIABPurposeAccepted',() async {
+                  var val = await TCConsentAPI.isIABPurposeAccepted(1);
+                  print("isIABPurposeAccepted = ${val}");
+                }),
+                buildTextButton('isVendorAccepted',() async {
+                  var val = await TCConsentAPI.isVendorAccepted(1);
+                  print("isVendorAccepted = ${val}");
+                }),
+                buildTextButton('isCategoryAccepted',() async {
+                  var val = await TCConsentAPI.isCategoryAccepted(1);
+                  print("isCategoryAccepted = ${val}");
+                }),
+                buildTextButton('getLastTimeConsentWasSaved',() async {
+                  var val = await TCConsentAPI.getLastTimeConsentWasSaved();
+                  print("getLastTimeConsentWasSaved = ${val}");
+                }),
+
+                buildTextButton('isConsentAlreadyGiven',() async {
+                  var val = await TCConsentAPI.isConsentAlreadyGiven();
+                  print("isConsentAlreadyGiven = ${val}");
+                }),
+
 
               ],
             ),