+++ ios/Classes/TCConsentPlugin.h
@@ -1,5 +1,6 @@
 #import <Flutter/Flutter.h>
 #import <TCCore/TCUser.h>
+#import <TCCore/TCLogger.h>
 #if __has_include(<TCConsent_IAB/TCPrivacyCallbacks.h>)
 #import <TCConsent_IAB/TCPrivacyCallbacks.h>
 #else